preparations for the implementations of DP and MN surfaces

+++ SRC/material/nD/DataDrivenMaterials/DruckerPragerDMM.h
@@ -30,7 +30,7 @@ public:
 	DruckerPragerDMM(int tag, double rho,
 		double Kref, double Gref, double Pref, double modn,
 		std::shared_ptr<DataDrivenNestedSurfaces> theData,
-		int dataDriverType, int integrationType, bool verbosity);
+		double dataDriverType, int integrationType, bool verbosity);
 
 	//null constructor
 	DruckerPragerDMM();
