+++ SRC/material/nD/DataDrivenMaterials/VonMisesDMM.cpp
@@ -36,7 +36,7 @@
 namespace global {
 	class MaterialList {
 	public:
-		int size = 0;
+		size_t size = 0;
 		int next_tag = 0;
 		std::vector<VonMisesDMM*> mptr;
 
@@ -76,7 +76,7 @@ auto VonMiseslist = global::MaterialList();
 VonMisesDMM::VonMisesDMM(int tag, double r0,
 	double K0, double G0, double P0, double m0,
 	std::shared_ptr<DataDrivenNestedSurfaces> ys,
-	int ddtype, int itype, bool verbosity)
+	double ddtype, int itype, bool verbosity)
 	:MultiYieldSurfaceHardeningSoftening(tag, ND_TAG_VonMisesDMM, r0,
 		K0, G0, P0, m0, ys, ddtype, itype, verbosity)
 {
