+++ SRC/material/nD/DataDrivenMaterials/VonMisesDMM.h
@@ -108,7 +108,7 @@ public:
 	VonMisesDMM(int tag, double rho,
 		double Kref, double Gref, double Pref, double modn,
 		std::shared_ptr<DataDrivenNestedSurfaces> theData,
-		int dataDriverType, int integrationType, bool verbosity);
+		double dataDriverType, int integrationType, bool verbosity);
 
 	//null constructor
 	VonMisesDMM();
